@@ -9,58 +9,64 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "bundle-report": "ng build --prod --stats-json && webpack-bundle-analyzer dist/stats.json"
+    "bundle-report": "ng build --prod --stats-json && webpack-bundle-analyzer dist/stats.json",
+    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
+    "serve:ssr": "node dist/server",
+    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
+    "build:client-and-server-bundles": "ng build --prod && ng run carloyi-pwa:server:production"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.7",
-    "@angular/cdk": "^7.3.3",
-    "@angular/common": "^7.2.7",
-    "@angular/compiler": "^7.2.7",
-    "@angular/core": "^7.2.7",
-    "@angular/fire": "^5.1.1",
-    "@angular/forms": "^7.2.7",
-    "@angular/http": "^7.2.7",
-    "@angular/material": "^7.3.3",
-    "@angular/material-moment-adapter": "^7.3.3",
-    "@angular/platform-browser": "^7.2.7",
-    "@angular/platform-browser-dynamic": "^7.2.7",
-    "@angular/platform-server": "^7.2.7",
-    "@angular/platform-webworker": "^7.2.11",
-    "@angular/platform-webworker-dynamic": "^7.2.11",
-    "@angular/router": "^7.2.7",
-    "@angular/service-worker": "^7.2.7",
-    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@angular/animations": "^8.2.5",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^8.2.5",
+    "@angular/compiler": "^8.2.5",
+    "@angular/core": "^8.2.5",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "^8.2.5",
+    "@angular/material": "^7.3.7",
+    "@angular/material-moment-adapter": "^7.3.7",
+    "@angular/platform-browser": "^8.2.5",
+    "@angular/platform-browser-dynamic": "^8.2.5",
+    "@angular/platform-server": "^8.2.5",
+    "@angular/platform-webworker": "^8.2.5",
+    "@angular/platform-webworker-dynamic": "^8.2.5",
+    "@angular/router": "^8.2.5",
+    "@angular/service-worker": "^8.2.5",
+    "@ng-bootstrap/ng-bootstrap": "^4.2.2",
     "@ngu/carousel": "^1.5.5",
     "@nguniversal/express-engine": "^7.1.1",
+    "@nguniversal/module-map-ngfactory-loader": "^8.1.1",
     "@ngx-loading-bar/core": "^2.2.0",
     "@ngx-loading-bar/http": "^2.2.0",
     "@ngx-loading-bar/http-client": "^2.2.0",
     "@ngx-loading-bar/router": "^2.2.0",
     "@ngx-meta/core": "^7.0.0",
-    "@ngx-progressbar/core": "^5.1.2",
-    "@types/express": "^4.16.0",
-    "@types/socket.io": "^1.4.36",
-    "angular": "^1.7.7",
-    "angular-animate": "^1.7.7",
-    "angular-aria": "^1.7.7",
-    "angular-bootstrap-md": "^7.4.4",
+    "@ngx-progressbar/core": "^5.3.2",
+    "@types/express": "^4.17.1",
+    "@types/socket.io": "^1.4.40",
+    "angular": "^1.7.8",
+    "angular-animate": "^1.7.8",
+    "angular-aria": "^1.7.8",
+    "angular-bootstrap-md": "^7.5.4",
     "angular-font-awesome": "^3.1.2",
     "angular-loading-bar": "^0.9.0",
-    "angular-material": "^1.1.13",
+    "angular-material": "^1.1.20",
     "angular-material-prefix-updater": "0.0.6",
-    "angular-messages": "^1.4.14",
+    "angular-messages": "^1.7.8",
     "angular-svg": "^2.0.8",
     "angular-webstorage-service": "^1.0.2",
     "angular2-image-zoom": "^1.2.1",
-    "angularfire2": "^5.0.0-rc.9",
-    "animate.css": "^3.6.1",
+    "angularfire2": "^5.2.1",
+    "animate.css": "^3.7.2",
     "autoprefixer": "^8.2.0",
     "bootstrap": "4.0.0-beta.2",
-    "chart.js": "^2.5.0",
-    "core-js": "^2.4.1",
-    "express": "^4.16.3",
-    "firebase": "^5.8.2",
+    "chart.js": "^2.8.0",
+    "codemirror": "^5.49.0",
+    "core-js": "^2.6.9",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "firebase": "^6.6.0",
     "font-awesome": "^4.7.0",
     "geolocation": "^0.2.0",
     "hamburgers": "^0.9.3",
@@ -68,59 +74,64 @@
     "jquery": "^3.4.1",
     "jsonfile": "^4.0.0",
     "material-design-icons": "^3.0.1",
-    "mdbootstrap": "^4.5.6",
+    "mdbootstrap": "^4.8.9",
+    "mock-browser": "^0.92.14",
     "moment": "^2.24.0",
     "ng-animate": "^0.3.3",
-    "ng-inline-svg": "^8.0.1",
+    "ng-inline-svg": "^8.6.0",
     "ng-loading-bar": "0.0.10",
     "ng-socket-io": "^0.2.4",
-    "ng2-animate-on-scroll": "^1.0.8",
+    "ng2-animate-on-scroll": "^1.1.0",
     "ng2-ion-range-slider": "^2.0.0",
-    "ng2-pdf-viewer": "^5.2.3",
+    "ng2-pdf-viewer": "^5.3.4",
     "ngx-bootstrap": "^2.0.3",
-    "ngx-content-loading": "^0.1.3",
     "ngx-lightbox": "^1.2.0",
-    "ngx-loading": "^1.0.14",
-    "ngx-material-timepicker": "^2.8.3",
-    "ngx-pagination": "^3.1.1",
+    "ngx-loading": "^1.0.15",
+    "ngx-material-timepicker": "^2.13.0",
+    "ngx-pagination": "^3.3.1",
     "ngx-progressbar": "^2.1.1",
-    "ngx-toastr": "^9.1.2",
+    "ngx-toastr": "^9.2.0",
     "nouislider": "^11.1.0",
-    "placeholder-loading": "^0.1.4",
-    "popper.js": "^1.14.2",
+    "placeholder-loading": "^0.1.5",
+    "popper.js": "^1.15.0",
     "promise-polyfill": "8.1.0",
-    "rxjs": "^6.4.0",
-    "rxjs-compat": "^6.3.2",
-    "socket.io": "^2.1.1",
-    "sweetalert2": "^7.19.3",
-    "tslib": "^1.9.0",
-    "web-push": "^3.3.2",
-    "zone.js": "^0.8.29"
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.3",
+    "socket.io": "^2.2.0",
+    "sweetalert2": "^7.33.1",
+    "ts-loader": "^4.0.0",
+    "tslib": "^1.10.0",
+    "web-push": "^3.3.5",
+    "webpack-node-externals": "^1.7.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.802.2",
-    "@angular/cli": "^7.3.3",
-    "@angular/compiler-cli": "^7.2.7",
-    "@angular/language-service": "^7.2.7",
-    "@fortawesome/fontawesome-free": "^5.6.3",
-    "@types/bootstrap": "^4.1.2",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/jquery": "^3.3.22",
-    "@types/node": "^6.14.2",
-    "codelyzer": "^4.0.1",
+    "@angular-devkit/build-angular": "~0.803.3",
+    "@angular/cli": "^8.3.3",
+    "@angular/compiler-cli": "^8.2.5",
+    "@angular/language-service": "^8.2.5",
+    "@fortawesome/fontawesome-free": "^5.10.2",
+    "@types/bootstrap": "^4.3.1",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/jquery": "^3.3.31",
+    "@types/node": "^6.14.7",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.0.0",
+    "karma": "^3.1.4",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.1",
+    "localstorage-polyfill": "^1.0.1",
+    "protractor": "^5.4.2",
     "t-scroll": "^1.1.9",
+    "ts-loader": "^5.2.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.2.4",
-    "webpack-bundle-analyzer": "^3.4.1"
+    "typescript": "3.5.3",
+    "webpack-bundle-analyzer": "^3.4.1",
+    "webpack-cli": "^3.1.0"
   }
 }
